Add MVP EigenDA Support
Features:
* Supports writing framesets to EigenDA, and then posting frameset EigenDA "references"
  to L1 calldata.
* If the write of a frameset to EigenDA fails, then the original
  frameset is written to L1 calldata. This graceful fallback is
  supported by a new wrapper type that encapsulates the union type between
  raw frameset data and a frameset reference, and is implemented using
  protocol buffers.
* Reading is transparently handled on the derivation pipeline side.

Caveats:
* The introduction of the concept of external data availability probably
  demands a refactor and reorganization of some logic. This changeset
  does not attempt any of this, instead inserting hacky logic in sometimes
  undesirable places to achieve this the intended proof-of-concept goal.
* op-programs in now broken an needs update to support "off-chain" DA.
* op-e2e is also broken and will need fixing before end-to-end tests
  can be run successfully.

+++ bedrock-devnet/devnet/__init__.py
@@ -35,7 +35,7 @@ class ChildProcess:
         try:
             func(*args)
         except Exception as e:
-            self.errq.put(str(e))
+            self.errq.put(e)
 
     def start(self):
         self.process.start()
@@ -109,6 +109,7 @@ def deploy_contracts(paths):
     res = eth_accounts('127.0.0.1:8545')
 
     response = json.loads(res)
+    print(response)
     account = response['result'][0]
     log.info(f'Deploying with {account}')
 
@@ -132,7 +133,10 @@ def deploy_contracts(paths):
         '--unlocked'
     ], env={}, cwd=paths.contracts_bedrock_dir)
 
-    shutil.copy(paths.l1_deployments_path, paths.addresses_json_path)
+    if os.path.exists(paths.l1_deployments_path):
+      shutil.copy(paths.l1_deployments_path, paths.addresses_json_path)
+    else:
+        raise Exception(f"File {paths.l1_deployments_path} does not exist for copy")
 
     log.info('Syncing contracts.')
     run_command([
@@ -161,7 +165,10 @@ def devnet_l1_genesis(paths):
     forge.join()
     err = forge.get_error()
     if err:
-        raise Exception(f"Exception occurred in child process: {err}")
+      try:
+        raise err
+      except subprocess.CalledProcessError as calledProcessErr:
+        raise Exception(f"Exception occurred in child process: {calledProcessErr}, with logs {calledProcessErr.stderr}")
 
     res = debug_dumpBlock('127.0.0.1:8545')
     response = json.loads(res)
@@ -304,9 +311,9 @@ def devnet_test(paths):
 
 def run_command(args, check=True, shell=False, cwd=None, env=None, timeout=None):
     env = env if env else {}
-    return subprocess.run(
+    completed_process = subprocess.run(
         args,
-        check=check,
+        check=False,  # Handle the check manually for custom error messages
         shell=shell,
         env={
             **os.environ,
@@ -315,6 +322,17 @@ def run_command(args, check=True, shell=False, cwd=None, env=None, timeout=None)
         cwd=cwd,
         timeout=timeout
     )
+
+    # Check if the subprocess was successful
+    if check and completed_process.returncode != 0:
+        # Raise an exception with stderr as the error message
+        raise subprocess.CalledProcessError(
+            completed_process.returncode,
+            completed_process.args,
+            stderr=completed_process.stderr.strip()
+        )
+
+    return completed_process
 
 
 def wait_up(port, retries=10, wait_secs=1):
